pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vaultchef"
version = "0.1.0"
description = "Terminal-only cookbook builder for Obsidian vaults"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [
  { name = "Vaultchef" }
]
dependencies = [
  "PyYAML>=6.0",
  "textual>=0.80",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "ruff>=0.6",
  "mypy>=1.11",
]

[project.scripts]
vaultchef = "vaultchef.cli:main"

[tool.pytest.ini_options]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
warn_unused_configs = true
disable_error_code = ["import-untyped"]

[[tool.mypy.overrides]]
module = ["vaultchef.tui", "vaultchef.tui.*"]
ignore_errors = true
